Refactor code into `ConvertToCSVService`
