+++ app/services/streams/stream_transform_csv.service.js
@@ -5,6 +5,7 @@
  */
 
 const { Transform } = require('stream')
+const ConvertToCSVService = require('../convert_to_csv.service')
 
 /**
  * Returns a Transform stream which turns an incoming array into a CSV row, ie. comma-separated, with each element
@@ -17,22 +18,7 @@ class StreamTransformCSVService {
     return new Transform({
       objectMode: true,
       transform: function (array, _encoding, callback) {
-        const wrapped = array.map(element => {
-          // JSON objects give us `[object Object]` when we use them in a template literal. We want to stringify these
-          // objects to prevent this, but we ONLY want to stringify these types of objects
-          if (`${element}` === '[object Object]') {
-            element = JSON.stringify(element)
-          }
-          // If element is a string then convert all quotes " to two quotes "" as required by CSV format
-          if (typeof element === 'string') {
-            element = element.replace(/"/g, '""')
-          }
-          // Finally, return element wrapped in quotes
-          return `"${element}"`
-        })
-        const csv = wrapped.join()
-
-        callback(null, csv.concat('\n'))
+        callback(null, ConvertToCSVService.go(array))
       }
     })
   }
